Convert API tests to new test fixtures
Fix fixture name conflict. Add missing test_routes.py file
from previous commits.

Affects: https://www.notion.so/lyrasis/Refactor-API-test-suite-9852f7a8e4484cc6bed8235acb20a7fe
