+++ tests/api/lcp/conftest.py
@@ -1,0 +1,3 @@
+# Pull in the session_fixture defined in core/testing.py
+# which does the database setup and initialization
+pytest_plugins = ["core.testing"]
