+++ tests/fixtures/odl.py
@@ -1,14 +1,24 @@
 import json
 import types
-from typing import Callable, Tuple, Any, Optional
+from typing import Any, Callable, Optional, Tuple
 
 import pytest
 from _pytest.monkeypatch import MonkeyPatch
 
 from api.odl import ODLAPI
-from core.model import get_one_or_create, Collection, Patron, LicensePool, Loan, Representation, Library, Work, License, \
-    IntegrationClient
-from core.testing import MockRequestsResponse, DatabaseTest
+from core.model import (
+    Collection,
+    IntegrationClient,
+    Library,
+    License,
+    LicensePool,
+    Loan,
+    Patron,
+    Representation,
+    Work,
+    get_one_or_create,
+)
+from core.testing import DatabaseTest, MockRequestsResponse
 from core.util.http import HTTP
 from tests.fixtures.api_odl_files import ODLAPIFilesFixture
 from tests.fixtures.database import DatabaseTransactionFixture
